Cleanup of sparse bus complete (+-)
Dense bus and related are still present, but in a future
I'd like to remove them and use only sparse bus (dictionarry based)
representation only. It will be a bit slower and more memory consuming,
but it will be easier to understand for now. When we get full support
we might add another versions too.

Dense bus string output will still be possible using _incremment_addr().

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>